Trait reactive_graph_dynamic_graph_api::dynamic_graph_schema_manager::DynamicGraphSchemaManager

pub trait DynamicGraphSchemaManager:
    Send
    + Sync
    + Lifecycle {
    // Required methods
    fn is_type_system_modified(&self) -> bool;
    fn get_schema_builder_context(&self) -> SchemaBuilderContext;
    fn get_schema_builder(&self) -> SchemaBuilder;
    fn create_dynamic_schema<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Schema, SchemaError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn regenerate_dynamic_schema<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SchemaError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn regenerate_dynamic_schema_if_modified<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SchemaError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_dynamic_schema<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Schema>, SchemaError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn is_type_system_modified(&self) -> bool

Returns true, if the type system has been modified.

fn get_schema_builder_context(&self) -> SchemaBuilderContext

Returns a new schema builder context.

fn get_schema_builder(&self) -> SchemaBuilder

Returns a new schema builder.

fn create_dynamic_schema<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Schema, SchemaError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the SDL of the Dynamic Graph Schema.

fn regenerate_dynamic_schema<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(), SchemaError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Regenerates the Dynamic Graph Schema.

fn regenerate_dynamic_schema_if_modified<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(), SchemaError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Regenerates the Dynamic Graph Schema if and only if the type system has been modified.

fn get_dynamic_schema<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Arc<Schema>, SchemaError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the Dynamic Graph Schema.